jungseo_week10_hw_django

+++ post/views.py
@@ -1,55 +1,41 @@
-from django.shortcuts import render
-
-# Create your views here.
-from rest_framework.response import Response
-from .models import Post
-from rest_framework.decorators import api_view
-from django.views.decorators.csrf import csrf_exempt
-from .serializers import PostSerializer
-@csrf_exempt
-@api_view(['POST'])
-def CreatePostView(request):
-    """
-    request 를 보낼 때 post 의 title 과 content 를 보내야합니다.
-    """
-    title = request.data.get('title')
-    content = request.data.get('content')
-    post = Post.objects.create(title=title, content=content)
-    return Response({"msg":f"'{post.title}'이 생성되었어요!"})
-
-
-@api_view(['GET'])
-def ReadAllPostView(request):
-    posts = Post.objects.all()
-    contents = [{post.title:post.content} for post in posts]
-    return Response({"posts":contents})
-
-from rest_framework.views import APIView
 from rest_framework import status
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from django.db.models import Count
+from tag.models import Tag
+
+from .models import Like, Post
+from .serializers import PostSerializer
 
 
 class PostListView(APIView):
-
-		### 얘네가 class inner function 들! ###
-    def get(self, request): 
-        posts = Post.objects.all()
+    def get(self, request):
+        posts = Post.objects.all().annotate(like_count=Count('like_users')).order_by('-like_count')
         serializer = PostSerializer(posts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        title = request.data.get('title')
-        content = request.data.get('content')
+        author = request.user
+        title = request.data.get("title")
+        content = request.data.get("content")
+        tag_contents = request.data.get("tags")
+
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
         if not title or not content:
-            return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)
+            return Response({"detail": "[title, description] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
+	    
+        post = Post.objects.create(title=title, content=content, author=author)
+	    
+        for tag_content in tag_contents:
+            if not Tag.objects.filter(content=tag_content).exists():
+                post.tags.create(content=tag_content)
+            else:
+                post.tags.add(Tag.objects.get(content=tag_content))
+        
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-    
-
-
-    
-    
-
 
 class PostDetailView(APIView):
     def get(self, request, post_id):
@@ -57,30 +43,73 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_200_OK)
-        
+
     def delete(self, request, post_id):
         try:
             post = Post.objects.get(id=post_id)
         except:
-            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)    
+        
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+        
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
-    def put(self, request, post_id):
+    
+    def patch(self, request, post_id):
         try:
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
 
-        # post.content = request.data.get('content')
-        # post.title = request.data.get('title')
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+        serializer = PostSerializer(post, data=request.data, partial=True)
 
-        # post.save()
+        tag_contents = request.data.get("tags")
+
+        post.tags.clear()  # 처음에는 clear
+        for tag_content in tag_contents:
+            if not Tag.objects.filter(content=tag_content).exists():  
+                post.tags.create(content=tag_content)
+            else:
+                post.tags.add(Tag.objects.get(content=tag_content))
+
+        if not serializer.is_valid():
+            return Response({"detail": "data validation error"}, status=status.HTTP_400_BAD_REQUEST)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+# 7th week(ManyToMany Field_Like)
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
         
-        if serializer.is_valid():
-            serializer.save()
-            serializer = PostSerializer(post,data=request.data)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-    
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+        like_list = post.like_set.filter(user=request.user)
+
+        if like_list.count() > 0:
+            post.like_set.get(user=request.user).delete()
+        else:
+            Like.objects.create(user=request.user, post=post)
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+class MyPostListView(APIView):
+    def get(self, request):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        posts = Post.objects.filter(author=request.user)
+        serializer = PostSerializer(posts, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
